Fix missing turrets.
This bug was driving me crazy. I spend hours on it.
The turrets weren't missing. One one turret was allowed to be 'alive' at
once.

This is due to behavoir of oprators.
I swapped the expression
  a |= b
from dlang to
  a = a || b
in golang

Unfortunately this isn't exactly the same. || is a shortcut operator is
dlang. But |= in dlang is not. it still evaluates the right hand side
even if the left hand side is already true.